refactor: reportPage 책임 분리 및 empty state 연결
