+++ src/pages/reports/ReportsPage.css.ts
@@ -25,11 +25,6 @@ export const pageState = style({
   minHeight: '32rem',
 });
 
-export const emptyActions = style({
-  display: 'flex',
-  gap: vars.space.s3,
-});
-
 export const reportContent = style({
   display: 'flex',
   flexDirection: 'column',
